requirements.txt
dotenv>=1.0.1
sqlalchemy>=2.0.37
mysql-connector